Job start logic ignores finished jobs for 2 minutes.
Since the build bot is designed to be stateless and restartable at any point,
it does not know whether a new build has been requested.

Thus, it has no way of knowing whether a build reported by jenkins is the result
of the build it requested, or some older build.

To avoid reporting on stale finished jobs, we ignore finished builds
for 2 minutes after starting a new job, hoping that the newly queued
or building job will show up by then.