Don't know when to kill PR commenter; keep it alive
The job start watcher may call at any time as a result
of a PLS SYNCH or a PLS REBUILD ALL, so we can't stop the commenter
when the PR's status seems definitive, as we may want to force-update
the other statuses regardless.

Thus, just keep it alive as long as its parent is.
TODO: figure out how all these actors relate and make sure
to stop them when no longer needed